Fixed TaskCompletionSource.SetResult called more than once in YubikeyOtpValidator
The test host crashed in the event bus branch, maybe because of the much heavier thread pool usage there.
But this bug could possibly occur independently from the event bus tests. The issue was, that the `TaskCompletionSource.SetResult` method has been called multiple times and I fixed it by switching to `TrySetResult` (Line 42).

```
The active test run was aborted. Reason: Test host process crashed : Unhandled exception. System.InvalidOperationException: An attempt was made to transition a task to a final state when it had already completed.
   at System.Threading.Tasks.TaskCompletionSource`1.SetResult(TResult result)
   at Servus.Core.Security.Hardware.Yubikey.YubikeyOtpValidator.<>c__DisplayClass4_0.<<ValidateAsync>b__0>d.MoveNext() in D:\_\Servus.Core\src\Servus.Core\Security\Hardware\Yubikey\YubikeyOtpValidator.cs:line 49
--- End of stack trace from previous location where exception was thrown ---
   at System.Threading.Tasks.Task.<>c.<ThrowAsync>b__139_1(Object state)
   at System.Threading.QueueUserWorkItemCallbackDefaultContext.Execute()
   at System.Threading.ThreadPoolWorkQueue.Dispatch()
   at System.Threading._ThreadPoolWaitCallback.PerformWaitCallback()
```

I also made some tweaks to the async code and tests, but the logic is unchanged.